#1686 - Start Refactor 3, change storage type to long (bigint), use jsonb for most stat data, store game data in tables. Change some id's to snowflakes.

+++ core/src/main/java/me/mykindos/betterpvp/core/client/achievements/impl/general/deaths/DeathAchievement.java
@@ -37,7 +37,7 @@ public class DeathAchievement extends SingleSimpleAchievement {
         super("Death", key,
                 AchievementCategories.DEATH_TYPE,
                 AchievementType.PERIOD,
-                (double) goal,
+                (long) goal,
                 new GenericStat(
                 MinecraftStat.builder()
                         .statistic(Statistic.DEATHS)
